% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellbench}
\alias{cellbench}
\title{cellbench}
\format{
An object of SingleCellExperiment class with 895 cells and 2001 genes.
}
\source{
https://github.com/LuyiTian/sc_mixology
}
\usage{
data(cellbench)
}
\description{
A single-cell RNA-seq dataset adapted from \href{https://github.com/LuyiTian/sc_mixology}{sc_mixology}
}
\keyword{datasets}
