% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLimma.R
\name{doLimma}
\alias{doLimma}
\alias{doVoom}
\title{Computing (trend-)Limma cell identity genes}
\usage{
doLimma(exprsMat, cellTypes)

doVoom(exprsMat, cellTypes)
}
\arguments{
\item{exprsMat}{expression matrix where columns denote cells and rows denote genes}

\item{cellTypes}{vector of cell type labels}
}
\value{
Returns a list, each element corresponds to a cell-type.

Returns a list, each element corresponds to a cell-type.
}
\description{
exprsMat accepts various matrix objects, including DelayedArray and HDF5Array for
out-of-memory computations.

exprsMat accepts various matrix objects, including DelayedArray and HDF5Array for
out-of-memory computations.
}
\examples{
set.seed(1234)
n = 50 ## genes, rows
p = 100 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

limma_output = doLimma(exprsMat = exprsMat, cellTypes = cellTypes)
voom_output = doVoom(exprsMat = exprsMat, cellTypes = cellTypes)
}
