% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalues.R
\name{getPvalues}
\alias{getPvalues}
\title{Extract differential statistics p-values from DX methods}
\usage{
getPvalues(object, returnDF = TRUE)
}
\arguments{
\item{object}{a computed result from Cepo()}

\item{returnDF}{Whether return a data.frame, default to TRUE.}
}
\value{
If returnDF is set to TRUE, then returns a data frame. Otherwise returns a list.
}
\description{
Extract differential statistics p-values from DX methods
}
\examples{
set.seed(1234)
n = 50 ## genes, rows
p = 100 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

cepo_output = Cepo(exprsMat = exprsMat, cellTypes = cellTypes)
limma_output = doLimma(exprsMat = exprsMat, cellTypes = cellTypes)
voom_output = doVoom(exprsMat = exprsMat, cellTypes = cellTypes)
ttest_output = doTtest(exprsMat = exprsMat, cellTypes = cellTypes)
wilcoxon_output = doWilcoxon(exprsMat = exprsMat, cellTypes = cellTypes)

getPvalues(cepo_output)
getPvalues(limma_output)
getPvalues(voom_output)
getPvalues(ttest_output)
getPvalues(wilcoxon_output)
}
