% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Extract differential statistics from DX methods}
\usage{
getStats(object, returnDF = TRUE)
}
\arguments{
\item{object}{a computed result from Cepo()}

\item{returnDF}{Whether return a data.frame, default to TRUE.}
}
\value{
If returnDF is set to TRUE, then returns a data frame. Otherwise returns a list.
}
\description{
Extract differential statistics from DX methods
}
\examples{
set.seed(1234)
n = 50 ## genes, rows
p = 100 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

cepo_output = Cepo(exprsMat = exprsMat, cellTypes = cellTypes)
limma_output = doLimma(exprsMat = exprsMat, cellTypes = cellTypes)
voom_output = doVoom(exprsMat = exprsMat, cellTypes = cellTypes)

getStats(cepo_output)
getStats(limma_output)
getStats(voom_output)
}
