% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot densities}
\usage{
plot_densities(
  x,
  dx_list,
  n_genes = NULL,
  assay = "logcounts",
  celltype_col,
  celltype,
  genes = NULL,
  hist_plot = TRUE,
  color = NULL
)
}
\arguments{
\item{x}{a \code{\linkS4class{SummarizedExperiment}} or a \code{\linkS4class{SingleCellExperiment}} object.}

\item{dx_list}{a list containing the ordered output of Dx runs.}

\item{assay}{a character ("logcounts" by default),
indicating the name of the assays(x) element which stores the expression data (i.e., assays(x)$name_assays_expression).
We strongly encourage using normalized data, such as counts per million (CPM) or log-CPM.}

\item{celltype_col}{X.}

\item{celltype}{a character, indicating the name of the cluster to plot. Default is all clusters.}

\item{color}{a named color vector.}

\item{gene}{a character, indicating the name of the gene to plot. Default is NULL. If is.null(gene), this argument overrides the \code{dx_list}.}

\item{hist}{a logical, indicating whether to plot histogram (if TRUE) or not (if FALSE).}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot densities
}
\examples{
data("cellbench", package = "Cepo")
cellbench
ds_res <- Cepo(logcounts(cellbench), cellbench$celltype)

plot_densities(x = cellbench,
               dx_list = ds_res, 
               n_genes = 2,
               assay = "logcounts", 
               celltype_col = "celltype",
               celltype = names(ds_res))

plot_densities(x = cellbench,
               dx_list=ds_res,
               genes = unlist(lapply(ds_res, function(x) names(x)[10:11])),
               assay = "logcounts", 
               celltype_col = "celltype",
               celltype = names(ds_res))


}
\author{
Hani Jieun Kim \email{hani.kim@sydney.edu.au}

\code{plot_density} returns a \code{\link{ggplot}} object
with cell-type specific densities for a gene.
}
