% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_pancreas}
\alias{sce_pancreas}
\title{sce_pancreas}
\format{
An object of SingleCellExperiment class with 528 cells and 1358 genes.
}
\usage{
data(sce_pancreas)
}
\description{
A subsampled single-cell RNA-seq dataset
}
\keyword{datasets}
