% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cepo.R
\name{Cepo}
\alias{Cepo}
\title{Functions to generate differentially expressed genes}
\usage{
Cepo(exprsMat, cellTypes, exprs_pct = NULL)
}
\arguments{
\item{exprsMat}{expression matrix where columns denote cells and rows denote genes}

\item{cellTypes}{vector of cell type labels}

\item{exprs_pct}{Percentage of lowly expressed genes to remove. Default to NULL to not remove any genes.}
}
\value{
Returns a list of key genes.
}
\description{
Functions to generate differentially expressed genes
}
\examples{
set.seed(1234)
n = 100 ## genes, rows
p = 1000 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

Cepo(exprsMat = exprsMat, cellTypes = cellTypes)
}
