% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cepo.R
\name{Cepo}
\alias{Cepo}
\title{Functions to generate differentially expressed genes}
\usage{
Cepo(exprsMat, cellTypes, exprs_pct = 0.05, filter = FALSE)
}
\arguments{
\item{exprsMat}{expression matrix where columns denote cells and rows denote genes}

\item{cellTypes}{vector of cell type labels}
}
\description{
Functions to generate differentially expressed genes
}
\examples{

####### an example where `sce`` is a SingleCellExperiment object
mat <- logcounts(sce)
labels <- sce$lineage

ds_res <- Cepo(mat, labels, exprs_pct=0.05, filter=FALSE)

####### A quick simulation. Note that we need to have colnames and rownames on matrix ##########
set.seed(1234)
n = 1000 ## genes, rows
p = 10000 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)
mat_output = Cepo(exprsMat = exprsMat, cellTypes = cellTypes)

library(DelayedArray)
library(DelayedMatrixStats)
library(HDF5Array)
DelayedArray:::set_verbose_block_processing(TRUE)
## DelayedArray::setAutoBlockSize(size = 1e3)
## DelayedArray::getAutoBlockSize()
exprsMat_da = DelayedArray::DelayedArray(realize(exprsMat, "HDF5Array"))
class(seed(exprsMat_da))
da_output = Cepo(exprsMat = exprsMat_da, cellTypes = cellTypes)
identical(mat_output, da_output)

library(microbenchmark)
microbenchmark::microbenchmark(
mat_output = Cepo(exprsMat = exprsMat, cellTypes = cellTypes),
da_output = Cepo(exprsMat = exprsMat_da, cellTypes = cellTypes), 
times = 10)
}
