% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cepo.R
\name{Cepo}
\alias{Cepo}
\title{Computing Cepo cell identity genes}
\usage{
Cepo(
  exprsMat,
  cellTypes,
  exprsPct = NULL,
  computePvalue = NULL,
  variability = "CV",
  workers = 1L,
  ...
)
}
\arguments{
\item{exprsMat}{expression matrix where columns denote cells and rows denote genes}

\item{cellTypes}{vector of cell type labels}

\item{exprsPct}{Percentage of lowly expressed genes to remove. Default to NULL to not remove any genes.}

\item{computePvalue}{Whether to compute p-values using bootstrap test. Default to NULL to not make computations.
Set this to an integer to set the number of bootstraps needed (recommend to be at least 100).}

\item{variability}{a character indicating the stability measure (CV, IQR, MAD, SD). Default is set to CV.}

\item{workers}{Number of cores to use. Default to 1, which invokes \code{BiocParallel::SerialParam}.
For workers greater than 1, see the \code{workers} argument in \code{BiocParallel::MulticoreParam} and \code{BiocParallel::SnowParam}.}

\item{...}{Additional arguments passed to \code{BiocParallel::MulticoreParam} and \code{BiocParallel::SnowParam}.}
}
\value{
Returns a list of key genes.
}
\description{
exprsMat accepts various matrix objects, including DelayedArray and HDF5Array for
out-of-memory computations. See vignette.
}
\examples{
library(SingleCellExperiment)
data('cellbench', package = 'Cepo')
cellbench
cepoOutput <- Cepo(logcounts(cellbench), cellbench$celltype)
cepoOutput
}
