% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cepo.R
\name{Cepo}
\alias{Cepo}
\title{Computing Cepo cell identity genes}
\usage{
Cepo(
  exprsMat,
  cellTypes,
  exprs_pct = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{exprsMat}{expression matrix where columns denote cells and rows denote genes}

\item{cellTypes}{vector of cell type labels}

\item{exprs_pct}{Percentage of lowly expressed genes to remove. Default to NULL to not remove any genes.}

\item{BPPARAM}{A BiocParallelParam object specifying how parallelization should be performed.
Default to BiocParallel::SerialParam()}
}
\value{
Returns a list of key genes.
}
\description{
exprsMat accepts various matrix objects, including DelayedArray and HDF5Array for
out-of-memory computations.
}
\examples{
set.seed(1234)
n = 50 ## genes, rows
p = 100 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

Cepo(exprsMat = exprsMat, cellTypes = cellTypes)
}
